BackEnd/requirements.txt
absl-py==2.1.0
alabaster==1.0.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bibtexparser==1.4.2
billiard==4.2.1
build==1.2.2.post1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.9
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect==1.2.0
Deprecated==1.2.15
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
django-tenant-schemas==1.12.0
django-tenants==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-spectacular==0.28.0
et_xmlfile==2.0.0
fake-useragent==2.0.3
filelock==3.16.1
flatbuffers==24.3.25
flower==2.0.1
free_proxy==1.1.3
fsspec==2024.10.0
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
humanize==4.11.0
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.0
namex==0.0.8
numpy==2.2.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
phonenumbers==8.13.52
pillow==11.0.0
pip-tools==7.4.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.1
PyMuPDFb==1.24.10
pyparsing==3.2.0
pyproject_hooks==1.2.0
pyquery==2.0.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
roman==4.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scholarly==1.7.11
selenium==4.27.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
termcolor==2.5.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
wrapt==1.17.0
wsproto==1.2.0
